ref: rearrange solution folders (#12)
