+++ Keycloak.Net.Sdk/Clients/Contracts/ClientResponseDto.cs
@@ -1,4 +1,4 @@
-namespace Keycloak.Net.Sdk.Contracts.Responses;
+namespace Keycloak.Net.Sdk.Clients.Contracts;
 
 public sealed record ClientResponseDto()
 {
