+++ Keycloak.Net.Sdk/Clients/Contracts/CreateClientRequestDto.cs
@@ -1,4 +1,4 @@
-namespace Keycloak.Net.Sdk.Contracts.Requests;
+namespace Keycloak.Net.Sdk.Clients.Contracts;
 
 public record CreateClientRequestDto()
 {
