+++ Keycloak.Net.Sdk/KeycloakManagement.cs
@@ -1,4 +1,9 @@
-﻿using Keycloak.Net.Sdk.Contracts;
+﻿using Keycloak.Net.Sdk.Athentications.Contracts;
+using Keycloak.Net.Sdk.Clients.Contracts;
+using Keycloak.Net.Sdk.Contracts;
+using Keycloak.Net.Sdk.Realms;
+using Keycloak.Net.Sdk.Roles.Contracts;
+using Keycloak.Net.Sdk.Users.Contracts;
 
 namespace Keycloak.Net.Sdk;
