+++ Keycloak.Net.Sdk/Roles/Contracts/ClientRoleResponseDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Keycloak.Net.Sdk.Contracts.Responses;
+namespace Keycloak.Net.Sdk.Roles.Contracts;
 
 public sealed record ClientRoleResponseDto()
 {
