+++ Keycloak.Net.Sdk/Users/Contracts/IUserManagement.cs
@@ -1,7 +1,7 @@
-using Keycloak.Net.Sdk.Contracts.Requests;
+using Keycloak.Net.Sdk.Athentications.Contracts;
 using Keycloak.Net.Sdk.Contracts.Responses;
 
-namespace Keycloak.Net.Sdk.Contracts;
+namespace Keycloak.Net.Sdk.Users.Contracts;
 
 public interface IUserManagement
 {
