+++ Keycloak.Net.Sdk/Users/Contracts/SetUserPasswordRequestDto.cs
@@ -1,0 +1,8 @@
+namespace Keycloak.Net.Sdk.Users.Contracts;
+
+public sealed record SetUserPasswordRequestDto()
+{
+    public string Type => "password";
+    public string Value { get; init; }
+    public bool Temporary { get; init; }
+}
